pyproject.toml
[project]
name = "fastmail-tui"
version = "0.1.0"
description = "Privacy-focused, AI-enhanced TUI email client for Fastmail"
authors = [{name = "Philip S. Wright", email = "philip@alwaysunder.io"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["fastmail", "email", "tui", "jmap", "terminal", "ai"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Communications :: Email",
]

dependencies = [
  "textual>=3.0.0",
  "rich>=13.0.0",
  "jmapc>=0.2.23",
  "httpx>=0.27.0",
  "anthropic>=0.40.0",
  "keyring>=25.0.0",
  "cryptography>=43.0.0",
  "pyyaml>=6.0",
  "click>=8.1.0",
  "pydantic>=2.0.0",
  "markdownify>=0.11.0",
  "pyperclip>=1.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-mock>=3.12.0",
  "mypy>=1.8.0",
  "ruff>=0.2.0",
]

[project.scripts]
fastmail-tui = "fastmail_tui.main:main"
fm = "fastmail_tui.main:main"

[project.urls]
Homepage = "https://github.com/pdubbbbbs/fastmail-tui"
Repository = "https://github.com/pdubbbbbs/fastmail-tui"
Issues = "https://github.com/pdubbbbbs/fastmail-tui/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fastmail_tui"]

[tool.ruff]
line-length = 100
indent-width = 2
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
